Moves setErrors to the node itself
