+++ packages/core/src/store.ts
@@ -369,6 +369,51 @@ export function applyMessages(
 }
 
 /**
+ * Error messages.
+ * @internal
+ */
+export type ErrorMessages = string[] | Record<string, string | string[]>
+
+/**
+ * Creates an array of message arrays from strings.
+ * @param errors - Arrays or objects of form errors or input errors
+ * @returns
+ * @internal
+ */
+export function createMessages(
+  node: FormKitNode,
+  ...errors: Array<ErrorMessages | undefined>
+): Array<FormKitMessage[] | Record<string, FormKitMessage[]>> {
+  return errors
+    .filter((m) => !!m)
+    .map((errorSet): FormKitMessage[] | Record<string, FormKitMessage[]> => {
+      const sourceKey = `${node.name}-set`
+      const make = (error: string) =>
+        createMessage({
+          key: error,
+          type: 'error',
+          value: error,
+          meta: { source: sourceKey },
+        })
+      if (Array.isArray(errorSet)) {
+        return errorSet.map((error) => make(error))
+      } else {
+        const errors: Record<string, FormKitMessage[]> = {}
+        for (const key in errorSet) {
+          if (Array.isArray(errorSet[key])) {
+            errors[key] = (errorSet[key] as string[]).map((error) =>
+              make(error)
+            )
+          } else {
+            errors[key] = [make(errorSet[key] as string)]
+          }
+        }
+        return errors
+      }
+    })
+}
+
+/**
  *
  * @param store - The store to apply this missed applications.
  * @param address - The address that was missed (a node path that didn't yet exist)
